+++ DatabaseApi/Models/MappableModel.cs
@@ -7,7 +7,7 @@ namespace DatabaseApi.Models
 {
     public class MappableModel
     {
-        public BigInteger Id { get; set; }
+        public int Id { get; set; }
         public string Name { get; set; }
         public BigInteger? WebId { get; set; }
         public string WebName { get; set; }
